// src/types/POIData.ts
export interface POIData {
  id: number;
  poi_name: string;
  poi_type: string;
  poi_type_list: string;
  poi_code_secondary: string;
  poi_code_old: string;
  address: string;
  latitude: number;
  longitude: number;
  city: string;
  district: string;
  district_code: string;
  state: string;
  country: string;
  h3index: string;
  google_plus_code: string;
  google_category_tags: string;
  gmaps_url: string;
  brands: string;
  location_status: string;
  polygon: string;
  polygon_prefix: string;
  polygon_area_sqm: number;
  polygon_shared_or_independent: string;
  population_density: number;
  population_density_bucket_id: string;
  number_of_pois_in_polygon: number;
  placeids_within: string;
  placeids_within_count: number;
  distance_between_point_and_building_centroid: number;
  distance_between_point_and_building_polygon: number;
  default_kring: number;
  visits_avg_daily: number;
  devices_avg_daily: number;
  devices_monthly: number;
  average_visit_duration: number;
  average_monthly_visitors: number;
  version: string;
  search_string: string;
}